Реализован просмотр пользователем состав своих команд

+++ common/models/TeamOptions.php
@@ -3,6 +3,7 @@
 namespace common\models;
 
 
+use common\models\search\TeamsSearch;
 use common\models\search\UserSearchNotInTeam;
 use common\models\tables\Teams;
 use common\models\tables\Users;
@@ -12,43 +13,69 @@ use yii\db\ActiveQuery;
 class TeamOptions extends Model
 {
     public $id;
-    protected $userId;
+    protected $_userId;
     
-    public function getTeam(){
-        return Teams::findOne($this->id);
+    public function rules() {
+        return [
+            [['id'], 'integer', 'min' => 1]
+        ];
+    }
+    
+    public function getTeam() {
+        if ($this->validate()) {
+            return Teams::findOne($this->id);
+        }
+        return null;
     }
     
     public function deleteUser($userId) {
-        $this->userId = $userId;
-        $team = Teams::find()
-            ->from('teams t')
-            ->where(['t.id' => $this->id])
-            ->innerJoinWith(['users u' => function ($query) {
-                /* @var ActiveQuery $query */
-                $query->onCondition(['u.id' => $this->userId]);
-            }])
-            ->andWhere(['u.id' => $userId])
-            ->one();
         
-        if (!is_null($team)) {
-            $team->unlink('users', $team->users[0], true);
+        if ($this->validate()) {
+            $this->_userId = $userId;
+            $team = Teams::find()
+                ->from('teams t')
+                ->where(['t.id' => $this->id])
+                ->innerJoinWith(['users u' => function ($query) {
+                    /* @var ActiveQuery $query */
+                    $query->onCondition(['u.id' => $this->_userId]);
+                }])
+                ->andWhere(['u.id' => $userId])
+                ->one();
+            if (!is_null($team)) {
+                $team->unlink('users', $team->users[0], true);
+            }
+            return $team;
         }
-        return $team;
+        return null;
     }
     
-    public function addUsers(array $usersId){
-        $users = Users::findAll($usersId);
-    
-        if (count($users) > 0) {
-            $team = Teams::findOne($this->id);
-            foreach ($users as $user) {
-                $team->link('users', $user);
+    public function addUsers(array $usersId) {
+        
+        if ($this->validate()) {
+            $users = Users::findAll($usersId);
+            if (count($users) > 0) {
+                $team = Teams::findOne($this->id);
+                foreach ($users as $user) {
+                    $team->link('users', $user);
+                }
             }
+            return count($users);
         }
-        return count($users);
+        return [];
     }
     
     public function getUsersSearchNotInTeam() {
-        return new UserSearchNotInTeam(['teamId' => $this->id]);
+        if ($this->validate()) {
+            return new UserSearchNotInTeam(['teamId' => $this->id]);
+        }
+        return null;
+    }
+    
+    public function getUserTeams($userId, $isDataProvider = false) {
+        $dataProvider = (new TeamsSearch(['userId' => $userId]))->search(null);
+        if ($isDataProvider) {
+            return $dataProvider;
+        }
+        return $dataProvider->query->all();
     }
 }
